requirements.txt
customtkinter
requests
pillow
